Big rubocop cleanup

+++ db/migrate/20250909132929_drop_audition_sessions.rb
@@ -1,10 +1,10 @@
+# frozen_string_literal: true
+
 class DropAuditionSessions < ActiveRecord::Migration[8.0]
   def change
-    if table_exists?(:auditions)
-      drop_table :auditions
-    end
-    if table_exists?(:audition_sessions)
-      drop_table :audition_sessions
-    end
+    drop_table :auditions if table_exists?(:auditions)
+    return unless table_exists?(:audition_sessions)
+
+    drop_table :audition_sessions
   end
 end
