+++ db/migrate/20250925133000_change_audition_session_audition_relationship.rb
@@ -1,9 +1,9 @@
+# frozen_string_literal: true
+
 class ChangeAuditionSessionAuditionRelationship < ActiveRecord::Migration[7.0]
   def change
     # Remove join table if it exists
-    if table_exists?(:audition_sessions_auditions)
-      drop_table :audition_sessions_auditions
-    end
+    drop_table :audition_sessions_auditions if table_exists?(:audition_sessions_auditions)
     # Add audition_session_id to auditions
     add_reference :auditions, :audition_session, foreign_key: true
   end
